Refactor EnumHelper and add extension methods
Refactored the `EnumHelper` class by moving the `GetDisplayName` method to a new `EnumExtension` class as an extension method for better modularity. Retained the `ToList<T>` method in `EnumHelper` with no changes.

Added a new `DateTimeExtension` class with a `ToVietnamTime` method to convert UTC `DateTime` to Vietnam time.

Introduced `EnumExtension` and `DateTimeExtension` under the `Project.Common.Extensions` namespace for better organization of extension methods.